Remove seqnum check without preliminary check of convolute identity in sorting algorithm
A comparison between seqnums does only make sense if the poems are part of the same convolute. Thus it is error-prone to generate a comparison string with date and seqnum without checking the convolute identity first.

The commit also removes a bug in the second if clause where only the values of the first element are compared with a lesser-than comparison (which can of course never be true since same values can't be different).